+++ src/components/project/ProjectView.svelte
@@ -90,7 +90,6 @@
     import {
         Restart,
         ShowKeyboardHelp,
-        ToggleBlocks as ToggleBlocks,
         VisibleModifyCommands,
         handleKeyCommand,
     } from '../editor/util/Commands';
@@ -115,6 +114,8 @@
         PROJECT_PARAM_EDIT,
         PROJECT_PARAM_PLAY,
     } from '../../routes/project/constants';
+    import Switch from '@components/widgets/Switch.svelte';
+    import { withVariationSelector } from '../../unicode/emoji';
 
     export let project: Project;
     export let original: Project | undefined = undefined;
@@ -1347,15 +1348,18 @@
                                         ></Toggle
                                     >
                                 {:else if tile.isSource()}
-                                    <Toggle
-                                        tips={$locales.get(
-                                            (l) => l.ui.source.toggle.blocks
+                                    <Switch
+                                        onLabel={withVariationSelector('🖱️')}
+                                        onTip={$locales.get(
+                                            (l) => l.ui.source.toggle.blocks.off
                                         )}
-                                        on={$blocks}
-                                        command={ToggleBlocks}
+                                        offLabel={withVariationSelector('⌨️')}
+                                        offTip={$locales.get(
+                                            (l) => l.ui.source.toggle.blocks.on
+                                        )}
                                         toggle={toggleBlocks}
-                                        >{ToggleBlocks.symbol}</Toggle
-                                    >
+                                        on={$blocks}
+                                    />
                                     <!-- Make a Button for every modify command -->
                                     {#each VisibleModifyCommands as command}<CommandButton
                                             {command}